Add googletest to the Drake superbuild.  Separate the Cascade, Feedback, and PD Control systems from System.h, and test each of them using googletest.
