+++ drake/examples/Pendulum/runDynamics.cpp
@@ -5,6 +5,7 @@
 #include "drake/systems/Simulation.h"
 #include "drake/systems/plants/BotVisualizer.h"
 #include "drake/systems/LCMSystem.h"
+#include "drake/systems/cascade_system.h"
 
 using namespace std;
 using namespace Drake;
