+++ drake/examples/Pendulum/runEnergyShaping.cpp
@@ -2,6 +2,8 @@
 #include "drake/examples/Pendulum/Pendulum.h"
 #include "drake/systems/Simulation.h"
 #include "drake/systems/plants/BotVisualizer.h"
+#include "drake/systems/cascade_system.h"
+#include "drake/systems/feedback_system.h"
 
 using namespace std;
 using namespace Drake;
