+++ drake/examples/Quadrotor/runDynamics.cpp
@@ -1,7 +1,10 @@
+
+#include <iostream>
 #include "drake/systems/plants/BotVisualizer.h"
-#include "drake/systems/plants/RigidBodySystem.h"
 #include "drake/systems/LCMSystem.h"
 #include "drake/systems/LinearSystem.h"
+#include "drake/systems/plants/RigidBodySystem.h"
+#include "drake/systems/cascade_system.h"
 
 #include "QuadrotorInput.h"
 #include "QuadrotorOutput.h"
