+++ drake/examples/Quadrotor/runLQR.cpp
@@ -1,8 +1,10 @@
 #include <iostream>
-#include "Quadrotor.h"
+#include "drake/examples/Quadrotor/Quadrotor.h"
 #include "drake/systems/Simulation.h"
 #include "drake/systems/controllers/LQR.h"
 #include "drake/systems/plants/BotVisualizer.h"
+#include "drake/systems/cascade_system.h"
+#include "drake/systems/feedback_system.h"
 
 using namespace std;
 using namespace Eigen;
